UPDATE: [CHORE Fixes #166253110] move specific routes to their routes
